requirements.txt
python-dotenv
langchain
openai
youtube-transcript-api
faiss-cpu
youtube_search
pytube